requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --no-emit-index-url --output-file=requirements.txt requirements.in
#
aiocontextvars==0.2.2
    # via -r requirements.in
aiodns==2.0.0
    # via -r requirements.in
aiohttp-devtools==0.13.1
    # via -r requirements.in
aiohttp-socks==0.5.5
    # via -r requirements.in
aiohttp==3.7.4
    # via
    #   -r requirements.in
    #   aiohttp-devtools
    #   aiohttp-socks
aiojobs==0.3.0
    # via -r requirements.in
anyio==3.4.0
    # via httpcore
asttokens==2.0.5
    # via devtools
async-timeout==3.0.1
    # via
    #   aiohttp
    #   python-socks
atoma==0.0.17
    # via -r requirements.in
attrs==21.4.0
    # via
    #   aiohttp
    #   aiohttp-socks
    #   atoma
beautifulsoup4==4.8.2
    # via
    #   -r requirements.in
    #   bs4
    #   pynliner
brotli==1.0.7
    # via -r requirements.in
bs4==0.0.1
    # via -r requirements.in
cachetools==5.3.2
    # via -r requirements.in
cchardet==2.1.6
    # via -r requirements.in
certifi==2020.6.20
    # via
    #   -r requirements.in
    #   httpcore
    #   httpx
    #   requests
cffi==1.15.0
    # via
    #   cryptography
    #   pycares
chardet==3.0.4
    # via
    #   -r requirements.in
    #   aiohttp
    #   readability-lxml
charset-normalizer==2.0.10
    # via requests
click==7.1.2
    # via
    #   -r requirements.in
    #   aiohttp-devtools
coreapi==2.3.3
    # via
    #   django-rest-swagger
    #   openapi-codec
coreschema==0.0.4
    # via coreapi
cryptography==3.3.2
    # via
    #   -r requirements.in
    #   pyjwt
    #   pyopenssl
cssselect==1.1.0
    # via readability-lxml
cssutils==2.3.0
    # via pynliner
decorator==5.1.0
    # via validators
defusedxml==0.7.1
    # via
    #   atoma
    #   python3-openid
devtools==0.8.0
    # via aiohttp-devtools
django-allauth==0.44.0
    # via -r requirements.in
django-debug-toolbar==2.2.1
    # via -r requirements.in
django-extensions==2.2.9
    # via -r requirements.in
django-optimistic-lock==1.0.0
    # via -r requirements.in
django-postgrespool2==1.0.1
    # via -r requirements.in
django-rest-auth==0.9.5
    # via -r requirements.in
django-rest-swagger==2.2.0
    # via -r requirements.in
django-seal==1.2.3
    # via -r requirements.in
django-timezone-field==4.0
    # via -r requirements.in
django==2.2.28
    # via
    #   -r requirements.in
    #   django-allauth
    #   django-debug-toolbar
    #   django-optimistic-lock
    #   django-postgrespool2
    #   django-rest-auth
    #   django-seal
    #   django-timezone-field
    #   djangorestframework
    #   jsonfield
djangorestframework==3.12.2
    # via
    #   -r requirements.in
    #   django-rest-auth
    #   django-rest-swagger
executing==0.8.2
    # via devtools
fake-useragent==0.1.11
    # via -r requirements.in
feedparser==6.0.11
    # via -r requirements.in
filelock==3.4.2
    # via tldextract
gunicorn==20.0.4
    # via -r requirements.in
h11==0.12.0
    # via httpcore
hashids==1.3.1
    # via -r requirements.in
html2text==2020.1.16
    # via -r requirements.in
httpcore==0.15.0
    # via httpx
httpx==0.23.0
    # via -r requirements.in
idna==3.3
    # via
    #   anyio
    #   requests
    #   rfc3986
    #   tldextract
    #   validr
    #   yarl
itypes==1.2.0
    # via coreapi
jinja2==3.0.3
    # via coreschema
jsonfield==3.1.0
    # via -r requirements.in
listparser==0.18
    # via -r requirements.in
loguru==0.3.1
    # via -r requirements.in
lxml==4.9.1
    # via
    #   -r requirements.in
    #   readability-lxml
lz4==3.1.0
    # via -r requirements.in
mako==1.2.2
    # via -r requirements.in
marisa-trie==1.1.0
    # via -r requirements.in
markdown==3.2.1
    # via -r requirements.in
markupsafe==2.0.1
    # via
    #   jinja2
    #   mako
multidict==5.2.0
    # via
    #   aiohttp
    #   yarl
netifaces==0.10.9
    # via -r requirements.in
oauthlib==3.2.1
    # via requests-oauthlib
openapi-codec==1.3.2
    # via django-rest-swagger
packaging==19.2
    # via -r requirements.in
psycopg2-binary==2.8.6
    # via -r requirements.in
pycares==4.2.0
    # via aiodns
pycparser==2.21
    # via cffi
pygments==2.11.1
    # via aiohttp-devtools
pyinstrument-cext==0.2.2
    # via
    #   -r requirements.in
    #   pyinstrument
pyinstrument==3.1.3
    # via -r requirements.in
pyjwt[crypto]==2.4.0
    # via
    #   -r requirements.in
    #   django-allauth
pynliner==0.8.0
    # via -r requirements.in
pyopenssl==19.1.0
    # via -r requirements.in
pyparsing==3.0.6
    # via
    #   packaging
    #   validr
pysocks==1.7.1
    # via -r requirements.in
python-dateutil==2.8.1
    # via
    #   -r requirements.in
    #   atoma
python-dotenv==0.12.0
    # via -r requirements.in
python-slugify==3.0.2
    # via -r requirements.in
python-socks[asyncio]==1.2.4
    # via
    #   -r requirements.in
    #   aiohttp-socks
python3-openid==3.2.0
    # via django-allauth
pytz==2021.3
    # via
    #   django
    #   django-timezone-field
pyyaml==5.4
    # via -r requirements.in
readability-lxml==0.8.1
    # via -r requirements.in
requests-file==1.5.1
    # via tldextract
requests-oauthlib==1.3.0
    # via django-allauth
requests==2.26.0
    # via
    #   -r requirements.in
    #   coreapi
    #   django-allauth
    #   requests-file
    #   requests-oauthlib
    #   tldextract
rfc3986[idna2008]==1.5.0
    # via httpx
sgmllib3k==1.0.0
    # via feedparser
simplejson==3.17.6
    # via django-rest-swagger
six==1.16.0
    # via
    #   asttokens
    #   cryptography
    #   django-extensions
    #   django-rest-auth
    #   packaging
    #   pyopenssl
    #   python-dateutil
    #   requests-file
    #   validators
sniffio==1.2.0
    # via
    #   anyio
    #   httpcore
    #   httpx
soupsieve==2.3.1
    # via beautifulsoup4
sqlalchemy==1.3.15
    # via
    #   -r requirements.in
    #   django-postgrespool2
sqlparse==0.4.2
    # via
    #   django
    #   django-debug-toolbar
supervisor==4.1.0
    # via -r requirements.in
terminaltables==3.1.10
    # via -r requirements.in
text-unidecode==1.2
    # via python-slugify
tldextract==5.1.1
    # via -r requirements.in
tqdm==4.44.1
    # via -r requirements.in
typing-extensions==4.0.1
    # via aiohttp
uritemplate==4.1.1
    # via coreapi
urllib3==1.26.6
    # via
    #   -r requirements.in
    #   requests
validators==0.14.0
    # via -r requirements.in
validr==1.2.2
    # via -r requirements.in
watchgod==0.7
    # via aiohttp-devtools
wcwidth==0.1.7
    # via -r requirements.in
whitenoise==5.0.1
    # via -r requirements.in
yarl==1.6.3
    # via
    #   -r requirements.in
    #   aiohttp

# The following packages are considered to be unsafe in a requirements file:
# setuptools
